Floor is now extended bei AbstractAisleComponents with stream implementation. Refactored return value of abstract methods too!

+++ src/tgw/warehouse/AbstractAisleComponents.java
@@ -2,10 +2,10 @@ package tgw.warehouse;
 
 public abstract class AbstractAisleComponents {
     public double getCurrentOccupancyRate() {
-        return 1.0 - calculateEmptyLocations()/ calculateLocationCapacity();
+        return 1.0 - (double) calculateEmptyLocations()/ (double) calculateLocationCapacity();
     }
 
-    protected abstract double calculateEmptyLocations();
+    protected abstract int calculateEmptyLocations();
     
-    protected abstract double calculateLocationCapacity();
+    protected abstract int calculateLocationCapacity();
 }
